chore: removing duplicate code
