+++ genomeGraph/search.go
@@ -18,91 +18,6 @@ import (
 	"github.com/vertgenlab/gonomics/numbers"
 )
 
-const (
-	defaultMatrixSize int  = 2480
-	leftTraversal     byte = 0
-	rightTraversal    byte = 1
-)
-
-type memoryPool struct {
-	Hits   []SeedDev
-	Worker []SeedDev
-}
-
-type MatrixAln struct {
-	m     [][]int64
-	trace [][]byte
-}
-
-type dnaPool struct {
-	Seq         []dna.Base
-	Path        []uint32
-	queryStart  int
-	queryEnd    int
-	targetStart int
-	targetEnd   int
-	currScore   int64
-}
-
-type dynamicScoreKeeper struct {
-	i        int
-	j        int
-	routeIdx int
-	currMax  int64
-	route    []cigar.ByteCigar
-}
-
-type scoreKeeper struct {
-	targetStart  int
-	targetEnd    int
-	queryStart   int
-	queryEnd     int
-	extension    int
-	currScore    int64
-	seedScore    int64
-	perfectScore int64
-	leftScore    int64
-	rightScore   int64
-	leftPath     []uint32
-	rightPath    []uint32
-	leftSeq      []dna.Base
-	rightSeq     []dna.Base
-	currSeq      []dna.Base
-	tailSeed     SeedDev
-
-	currSeed       SeedDev
-	leftAlignment  []cigar.ByteCigar
-	rightAlignment []cigar.ByteCigar
-}
-
-func NewDnaPool() sync.Pool {
-	return sync.Pool{
-		New: func() interface{} {
-			dnaSeq := dnaPool{
-				Seq:         make([]dna.Base, 0, 150),
-				Path:        make([]uint32, 0, 10),
-				queryStart:  0,
-				targetStart: 0,
-				targetEnd:   0,
-				queryEnd:    0,
-			}
-			return &dnaSeq
-		},
-	}
-}
-
-func NewMemSeedPool() sync.Pool {
-	return sync.Pool{
-		New: func() interface{} {
-			pool := memoryPool{
-				Hits:   make([]SeedDev, 0, 10000),
-				Worker: make([]SeedDev, 0, 10000),
-			}
-			return &pool
-		},
-	}
-}
-
 // AlignGraphTraversal performs graph traversal for alignment either to the left or right, calculating optimal alignment using dynamic programming.
 func AlignGraphTraversal(n *Node, seq []dna.Base, position int, currentPath []uint32, extension int, read []dna.Base, scores [][]int64, matrix *MatrixAln, sk *scoreKeeper, dynamicScore *dynamicScoreKeeper, pool *sync.Pool, direction byte) ([]cigar.ByteCigar, int64, int, int, []uint32) {
 	// Fetch a pooled object to reduce allocations.
